Address maintainer feedback: extract AdvancedView, add settings toggle, and revert icons
